requirements.txt
matplotlib==3.4.3
numpy==1.22.4
openai==0.27.5
pandas==1.2.3
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.9.3
torch==1.13.0
tqdm==4.64.1
transformers==4.12.5
